Condensed and moved contents of logging_files to core
